dashboard/requirements.txt
djangorestframework==3.7.7
